prideta apsauga

+++ NodeJs/2022-06-06/index.js
@@ -109,7 +109,11 @@ app.put('/update/:id', async (req, res) => {
 })  
 
 app.get('/prekes', (req, res) => {
-    res.sendFile(__dirname + '/templates/prekes.html')
+    if(req.session.loggedin) {
+        res.sendFile(__dirname + '/templates/prekes.html')
+    } else {
+        res.sendFile(__dirname + '/templates/login.html')
+    }
 })
 
 app.get('/prekes-all', async (req, res) => {
@@ -122,37 +126,33 @@ app.get('/prekes-all', async (req, res) => {
 })
 
 app.post('/prekes-new', async (req, res) => {
-    try {
+    if(req.session.loggedin) {
         await Daiktai.create(req.body)
         res.status(200).end()
-    } catch {
+    } else {
          res.sendStatus(500)
     }
 })
 
 app.delete('/prekes-delete/:id', async (req, res) => {
     const id = req.params.id
-    try {
+    if(req.session.loggedin) {
         await Daiktai.findByIdAndDelete(id)
         res.sendStatus(200)
-    } catch {    
+    } else {    
         res.sendStatus(500)
     }
 })  
 
 app.put('/prekes-update/:id', async (req, res) => {
     const id = req.params.id
-    console.log(req.body)
-    try {
+    if(req.session.loggedin) {
         await Daiktai.findByIdAndUpdate(id, req.body)
         res.sendStatus(200)
-    } catch {
+    } else {
         res.sendStatus(500)
     }
 })  
-
-
-
 
 
 app.listen(3000)
